Delete old boids and rename new boids file
