+++ boids.py
@@ -1,7 +1,9 @@
 #!/usr/bin/env python3
-from random import randint
+from random import randint, choice
 import pygame as pg
 import numpy as np
+import math
+import json
 
 '''
 PyNBoids - a Boids simulation - github.com/Nikorasu/PyNBoids
@@ -19,22 +21,28 @@ SHOWFPS = True          # Show frame rate
 MOUSEFEAR = True        # Is there a fear node on the cursor
 TUNING = {
     "max_speed": 150,       # Max movement speed
-    "max_force": 5,         # Max acceleration force
+    # "max_force": 5,         # Max acceleration force
     "weightings": {         # Force weightings
-        'sep': 1.8,
+        'sep': 1,
         'ali': 1,
-        'coh': 1,
-        'fear': 2
+        # 'coh': 1,
+        'decel': 1,
+        'fear': 2e6,
+        'wall': 2e7
     },
     "target_dist": 40,      # Target separation
     "influence_dist": {
         "boid": 200,        # "visibility" distance for the boids
-        "fear": 200
+        "fear": 200,
+        "wall": 25
     },
-    "fear_decay": 1,        # see below
-    "fear_const": .05,      # 1/(r/k)^a -> k is const, a is decay
-    "speed_decay": 2        # Decay rate of speed -> v /= speed_decay * dt
+    # "fear_decay": 1,        # see below
+    # "fear_const": .05,      # 1/(r/k)^a -> k is const, a is decay
+    # "speed_decay": 2        # Decay rate of speed -> v /= speed_decay * dt
+    "target_vel": pg.Vector2(0, 0)
 }
+PIX_PER_METER = 1.5     # Number of pixels per meter in the real world
+FEAR_SPEED = 300        # Speed of drones
 
 
 def clamp_magnitude(vector, magnitude):
@@ -48,207 +56,348 @@ def clamp_magnitude(vector, magnitude):
 
 
 class Boid(pg.sprite.Sprite):
-    def __init__(self, boidNum, data, drawSurf, cHSV=None):
+    def __init__(self, boidNum, data, render, drawSurf=None, cHSV=None):
         super().__init__()
 
         self.data = data # Stores positions & rotations of all boids and fears
         self.bnum = boidNum # This boid's number (ID)
-        self.drawSurf = drawSurf # Main screen surface
+        self.render = render # Whether to render the pygame screen or not
 
-        self.image = pg.Surface((15, 15)).convert() # Area to render boid onto
-        self.image.set_colorkey(0)
-        self.color = pg.Color(0)  # preps color so we can use hsva
-        if boidNum == 0:
-            self.color.hsva = (randint(0,360), 90, 90) if cHSV is None else cHSV # randint(5,55) #4goldfish
-        else:
-            self.color.hsva = (0, 0, 20)
-        pg.draw.polygon(self.image, self.color, ((7,0), (13,14), (7,11), (1,14), (7,0))) # Arrow shape
-        self.orig_image = pg.transform.rotate(self.image.copy(), -90)
+        self.ang = pg.Vector2(0,0)
+        self.accel = pg.Vector2(0,0)
+        self.vel = pg.Vector2(randint(-100, 100), randint(-100, 100))
+        self.pos = pg.Vector2(randint(50, WIDTH - 50), randint(50, HEIGHT - 50))
+        # Finally, output pos and vel to array
+        self.data.boids[self.bnum, :2] = [self.pos, self.vel]
 
-        self.bSize = 17 # "personal space" size
-        self.dir = pg.Vector2(1, 0)  # sets up forward direction
-        maxW, maxH = self.drawSurf.get_size()
-        self.rect = self.image.get_rect(center=(randint(50, maxW - 50), randint(50, maxH - 50)))
-        self.ang = randint(0, 360)  # random start angle, & position ^
-        self.vel = pg.Vector2(0,0)# pg.Vector2(randint(-100, 100), randint(-100, 100))
-        self.pos = pg.Vector2(self.rect.center)
+        if render:
+            self.drawSurf = drawSurf # Main screen surface
 
+            self.image = pg.Surface((15, 15)).convert() # Area to render boid onto
+            self.image.set_colorkey(0)
+            self.color = pg.Color(0)  # preps color so we can use hsva
+            if boidNum == 0:
+                self.color.hsva = (randint(0,360), 90, 90) if cHSV is None else cHSV # randint(5,55) #4goldfish
+            else:
+                self.color.hsva = (0, 0, 60)
+            pg.draw.polygon(self.image, self.color, ((7,0), (13,14), (7,11), (1,14), (7,0))) # Arrow shape
+            self.orig_image = pg.transform.rotate(self.image.copy(), -90)
+
+            # maxW, maxH = self.drawSurf.get_size()
+            self.rect = self.image.get_rect(center=(self.pos.x, self.pos.y))
+        
+
+    
     def update(self, dt, tuning):
-        max_speed = tuning["max_speed"]
-        max_force = tuning["max_force"]
+        def getNearest(type, num_select=7):
+            # Make list of nearby boids, sorted by distance
+            if type == "boid": objects = self.data.boids
+            elif type == "fear": objects = self.data.fears
+            elif type == "wall": objects = self.data.walls
+            otherBoids = np.delete(objects, self.bnum, 0) # delete self
+            array_dists = np.zeros(len(otherBoids))
+            for i, boid in enumerate(otherBoids):
+                if type == "boid": boid_pos = boid[0]
+                elif type == "fear" or type == "wall": boid_pos = boid
+                array_dists[i] = (self.pos - boid_pos).length() # Get list of distances
+            closeBoidIs = np.argsort(array_dists)[:num_select] # Look at $num_select$ closest boids only
+            neiboids = otherBoids[closeBoidIs] # pick out chosen boids
+            neiboids = neiboids[array_dists[closeBoidIs] < tuning["influence_dist"]["boid"]] # Select only boids within influence radius
 
+            return neiboids
 
-        def sep_function(distance):
-            y_int = 1 # y intercept
-            x_int = tuning["target_dist"] # x intercept
+        # Define vectors to hold unweighted acceleration values
+        sep = pg.Vector2(0, 0)
+        ali = pg.Vector2(0, 0)
+        decel = pg.Vector2(0, 0)
+        fear = pg.Vector2(0, 0)
+        wall = pg.Vector2(0, 0)
 
-            return - y_int / x_int * distance + y_int 
+        # Loop through near boids and sum sep and ali components
+        for boid in getNearest("boid"):
+            rel_pos = boid[0] - self.pos # r_{ij} in the paper
+            rel_vel = boid[1] - self.vel # v_j - v_i in the paper
 
-        def get_force(steer):
-            if steer.magnitude() != 0:
-                steer = steer.normalize()
-                steer *= max_speed
-                steer -= self.vel
-                steer = clamp_magnitude(steer, max_force)
-
-            return steer
-
-        target_dist = tuning["target_dist"] # Ideal separation from other boids
-        influence_dist = tuning["influence_dist"]
-
-        # Make list of nearby boids, sorted by distance
-        otherBoids = np.delete(self.data.boids, self.bnum, 0)
-        array_dists = (self.pos.x - otherBoids[:,0])**2 + (self.pos.y - otherBoids[:,1])**2 # distance^2 to save computation time
-        closeBoidIs = np.argsort(array_dists)[:7] # Look at 7 closest boids only
-        neiboids = otherBoids[closeBoidIs]
-        neiboids[:,4] = np.sqrt(array_dists[closeBoidIs])
-        neiboids = neiboids[neiboids[:,4] < influence_dist["boid"]]
-
-        # Initialise steering vectors
-        sep_steer = pg.Vector2(0,0)
-        ali_steer = pg.Vector2(0,0)
-        coh_steer = pg.Vector2(0,0)
-        fear_steer = pg.Vector2(0,0)
-
-        if neiboids.size > 1:  # if has neighbours, do math and sim rules
-            # Apply the 3 criteria - separation, cohesion, alignment
-
-            # Separation
-            sep_count = 0
-            for boid in neiboids:
-                if (boid[4] < target_dist):
-                    # Get normalised direction of force
-                    direction = (self.pos - pg.Vector2(boid[0:2].tolist())).normalize()
-                    # Weight by distance and add to sum
-                    sep_steer += direction * sep_function(boid[4])
-                    sep_count += 1
-            # Normalise by number of boids influencing
-            if sep_count > 0: sep_steer /= sep_count
-
-            # Alignment
-            ali_steer = pg.Vector2(np.mean(neiboids[:,2] / neiboids[:,4]), np.mean(neiboids[:,3] / neiboids[:,4]))
-
-
-            # Cohesion
-            # Weighted averages
-            # denom = np.sum(1 / neiboids[:,4])
-            # avg_pos_x = np.mean(neiboids[:,0] / neiboids[:,4]) / denom
-            # avg_pos_y = np.mean(neiboids[:,1] / neiboids[:,4]) / denom
-            # avg_pos = pg.Vector2(avg_pos_x, avg_pos_y)
-            # coh_steer = avg_pos - self.pos
-
-            avg_pos = (np.mean(neiboids[:,0]), np.mean(neiboids[:,1]))
-            coh_steer = pg.Vector2(avg_pos) - self.pos
+            sep += (1 - (tuning["target_dist"] / rel_pos.length())**3) * rel_pos
+            ali += rel_vel
         
-        # Fear
-        self.data.fears[:,2] = np.sqrt((self.pos.x - self.data.fears[:,0])**2 + (self.pos.y - self.data.fears[:,1])**2)
-        for fear in self.data.fears:
-            if (fear[2] < influence_dist["fear"]):
-                # Get normalised direction of force
-                direction = (self.pos - pg.Vector2(fear[0:2].tolist())).normalize()
-                # Weight by distance and add to sum
-                fear_steer += direction / (fear[2] / tuning["fear_const"])**tuning["fear_decay"]
-
-        # Get forces from steer vectors, including weightings
-        weightings = tuning["weightings"]
-        sep_force = get_force(sep_steer) * weightings["sep"]
-        ali_force = get_force(ali_steer) * weightings["ali"]
-        coh_force = get_force(coh_steer) * weightings["coh"]
-        fear_force = get_force(fear_steer) * weightings["fear"]
-
-        # Sum forces
-        force = sep_force + ali_force + coh_force + fear_force
-
-        # Adjust velocity to reflect force
-        self.vel += force
-
-        # Get angle of velocity for rendering
-        self.ang = self.vel.as_polar()[1]
-
-        # Adjusts angle of boid image to match heading
-        self.image = pg.transform.rotate(self.orig_image, -self.ang)
-        self.rect = self.image.get_rect(center=self.rect.center)  # recentering fix
-        self.dir = pg.Vector2(1, 0).rotate(self.ang).normalize()
+        # Calculate decel component
+        decel = tuning["target_vel"] - self.vel # v_d - v_i in the paper
         
-        # Ensure max_speed is not exceeded
-        self.vel = clamp_magnitude(self.vel, max_speed)
+        # Loop through fears and sum fear components
+        for fear_obj in self.data.fears:
+            rel_pos = self.pos - fear_obj # r_{pi} in the paper
 
-        # Add speed decay
-        self.vel /= tuning["speed_decay"] ** dt
+            if rel_pos.length() <= tuning["influence_dist"]["fear"]:
+                fear += (1 / rel_pos.length()**3) * rel_pos
+        
+        # Loop through walls and sum fear components
+        for wall_obj in self.data.walls:
+            diff = wall_obj[1] - wall_obj[0]
+            length2 = diff.length_squared()
 
-        if self.bnum == 0:
-            print(sep_force, ali_force, coh_force, fear_force)
-            #print(self.vel, self.vel.magnitude())
+            if length2 == 0: # End points of wall are identical
+                closest_point = wall_obj[0]
+            else:
+                # p = self.pos
+                # v = wall_obj[0]
+                # w = wall_obj[1]
+                # Consider the line extending the segment, parameterized as v + t (w - v).
+                # We find projection of point p onto the line. 
+                # It falls where t = [(p-v) . (w-v)] / |w-v|^2
+                # We clamp t from [0,1] to handle points outside the segment vw.
+                t = max(0, min(1, (self.pos - wall_obj[0]).dot(diff) / length2))
+                closest_point = wall_obj[0] + t * diff
 
-        # Change position based on velocity
+            rel_pos = self.pos - closest_point # r_{pi} in the paper
+
+            if rel_pos.length() <= tuning["influence_dist"]["wall"]:
+                if rel_pos.length() == 0: # Handle boid being exactly on the line
+                    rel_pos = pg.Vector2(choice([-2, -1, 1, 2]), choice([-2, -1, 1, 2]))
+                fear += (1 / rel_pos.length()**8) * rel_pos
+
+        # Apply weights
+        sep *= tuning["weightings"]['sep']
+        ali *= tuning["weightings"]['ali']
+        decel *= tuning["weightings"]['decel']
+        fear *= tuning["weightings"]['fear']
+        wall *= tuning["weightings"]['wall']
+
+        # Debug print output to help tune
+        # if self.bnum == 0:
+        #     print(sep, ali, decel, fear, sep="\t")
+
+        # Sum weighted components to get acceleration
+        self.accel = sep + ali + decel + fear + wall
+
+        # Change velocity and position based on acceleration
+        self.vel += self.accel * dt
+        self.vel = clamp_magnitude(self.vel, tuning["max_speed"])
         self.pos += self.vel * dt
 
-        # Update position of rendered boid
-        self.rect.center = self.pos
+        # Update data array
+        self.data.boids[self.bnum, :2] = [self.pos, self.vel]
+        
+        if self.render:
+            # Update position of rendered boid
+            self.rect.center = self.pos
 
-        # Finally, output pos and vel to array
-        self.data.boids[self.bnum,:4] = [self.pos[0], self.pos[1], self.vel[0], self.vel[1]]
+            # Get angle of velocity for rendering
+            self.ang = self.vel.as_polar()[1]
+
+            # Adjusts angle of rendered boid image to match heading
+            self.image = pg.transform.rotate(self.orig_image, -self.ang)
+            self.rect = self.image.get_rect(center=self.rect.center)  # recentering fix
 
 
 class Data():
-    def __init__(self, n_boids, n_fears=1):
-        self.boids = np.zeros((n_boids, 5), dtype=float)
-        self.fears = np.zeros((n_fears, 3), dtype=float)
+    def __init__(self, n_boids, n_fears=100, n_walls=100):
+        self.initBoids(n_boids)
+        self.initFears(n_fears, pg.Vector2(100,100))
+        self.initWalls(n_walls)
 
+    def initBoids(self, n_boids):
+        # Boids array
+        self.boids = np.empty((n_boids, 2), dtype=pg.Vector2)
+        self.boids.fill(pg.Vector2(0, 0))
 
-def main():
-    pg.init()  # prepare window
-    pg.display.set_caption("Sheeeeeeep") # Window title
+    def initFears(self, n_fears, start_pos=pg.Vector2(-10000, -10000)):
+        # Fear array
+        self.fears = np.empty((n_fears), dtype=pg.Vector2)
+        self.fears.fill(start_pos) # Initialise all fears far off screen
 
-    # setup fullscreen or window mode
-    if FLLSCRN:
-        currentRez = (pg.display.Info().current_w, pg.display.Info().current_h)
-        screen = pg.display.set_mode(currentRez, pg.SCALED)
-        pg.mouse.set_visible(False)
-    else: screen = pg.display.set_mode((WIDTH, HEIGHT), pg.RESIZABLE)
+        self.num_fears = 0
 
-    # If mouse controls fear
-    if MOUSEFEAR:
-        pg.mouse.set_visible(False)
+        # Fear target array
+        self.fear_targets = np.empty((n_fears), dtype=pg.Vector2)
+        self.fear_targets.fill(start_pos) # Initialise all fears' target position far off screen
+
+    def initWalls(self, n_walls):
+        # Walls array
+        self.walls = np.empty((n_walls, 2), dtype=pg.Vector2)
+        self.walls.fill(pg.Vector2(-10000, -10000)) # Initialise all walls far off screen
+
+        self.num_walls = 0
     
+    def makeWall(self, start_point, end_point):
+        self.walls[self.num_walls] = [start_point, end_point]
+        self.num_walls += 1
 
-    # Set up boids and their data
-    nBoids = pg.sprite.Group()
-    data = Data(BOIDZ)
-    for n in range(BOIDZ):
-        nBoids.add(Boid(n, data, screen))  # spawns desired # of boidz
+    def drawWalls(self, surface):
+        for wall in self.walls:
+            pg.draw.line(surface, (255, 0, 0), wall[0], wall[1], 5) # Draw line between points
+            pg.draw.circle(surface, (255, 0, 0), wall[0], 2.5) # Draw red circle on each end
+            pg.draw.circle(surface, (255, 0, 0), wall[1], 2.5) # Draw red circle on other end
 
-    clock = pg.time.Clock()
-    if SHOWFPS : font = pg.font.Font(None, 30)
+    def drawFears(self, surface):
+        for fear in self.fears:
+            pg.draw.circle(surface, (32, 0, 0), fear, TUNING["influence_dist"]["fear"]) # Draw red circle on mouse position
+        for fear in self.fears:
+            pg.draw.circle(surface, (255, 0, 0), fear, 5) # Draw red circle on mouse position
 
-    # main loop
-    while True:
-        # Get mouse position if MOUSEFEAR
-        if MOUSEFEAR:
-            mouse_pos = pg.mouse.get_pos()
-            
-            data.fears[0][0] = mouse_pos[0]
-            data.fears[0][1] = mouse_pos[1]
 
-        for e in pg.event.get():
-            # Handle quitting
-            if e.type == pg.QUIT or e.type == pg.KEYDOWN and e.key == pg.K_ESCAPE:
+class Simulation():
+    def __init__(self, num_fears, num_boids=50, render=True):
+        self.render = render # Whether to render the pygame screen or not
+
+        if self.render:
+            pg.init()  # prepare window
+            pg.display.set_caption("Sheeeeeeep") # Window title
+
+            # setup fullscreen or window mode
+            if FLLSCRN:
+                currentRez = (pg.display.Info().current_w, pg.display.Info().current_h)
+                self.screen = pg.display.set_mode(currentRez, pg.SCALED)
+                pg.mouse.set_visible(False)
+            else: self.screen = pg.display.set_mode((WIDTH, HEIGHT), pg.RESIZABLE)
+
+            # If mouse controls fear
+            if MOUSEFEAR:
+                pg.mouse.set_visible(False)
+
+            if SHOWFPS : self.font = pg.font.Font(None, 30)
+        
+        else:
+            self.screen = None
+        
+        self.clock = pg.time.Clock()
+
+        # Set up boids and their data
+        self.data = Data(num_boids, num_fears)
+
+        self.nBoids = pg.sprite.Group()
+        for n in range(num_boids):
+            self.nBoids.add(Boid(n, self.data, self.render, self.screen))  # spawns desired # of boidz
+    
+    def addWallsFromJSON(self, JSON):
+        num_segments = 0
+
+        min_point = pg.Vector2(
+            JSON[0]["points"][0][0],
+            JSON[0]["points"][0][1]
+        )
+        max_point = pg.Vector2(
+            JSON[0]["points"][0][0],
+            JSON[0]["points"][0][1]
+        )
+
+        for wall in JSON:
+            num_segments += len(wall["points"]) - 1
+
+            for point in wall["points"]:
+                min_point.x = min(min_point.x, point[0])
+                max_point.x = max(max_point.x, point[0])
+                min_point.y = min(min_point.y, point[1])
+                max_point.y = max(max_point.y, point[1])
+        
+        center_point = (min_point + max_point)/2
+
+        earth_circ = 40075e3 # km
+        m_per_lat = earth_circ / 360 # meters per degree of latitude
+        m_per_lng = earth_circ * math.cos(center_point.y / 180 * math.pi) / 360
+
+        def transformCoords(lat_long):
+            rel_lat_long = lat_long - min_point
+
+            return pg.Vector2(
+                rel_lat_long.y * m_per_lng * PIX_PER_METER,
+                HEIGHT - rel_lat_long.x * m_per_lat * PIX_PER_METER # Flipping becaue (0,0) is top left in pg
+            ) # switching x and y because lat and long are (y,x)
+
+        self.data.initWalls(num_segments)
+
+        print(transformCoords(pg.Vector2([
+                    51.62394653210728,
+                    -2.5112555701146744
+                ])))
+
+        for wall in JSON:
+            points = wall["points"]
+
+            for i in range(len(points) - 1):
+                self.data.makeWall(
+                    start_point=transformCoords(pg.Vector2(points[i][0], points[i][1])),
+                    end_point=transformCoords(pg.Vector2(points[i+1][0], points[i+1][1]))
+                )
+
+    def addTestWalls(self):
+        pad = 40
+
+        self.data.initWalls(7)
+
+        self.data.makeWall(pg.Vector2(pad, pad), pg.Vector2(WIDTH-pad, pad))
+        self.data.makeWall(pg.Vector2(WIDTH-pad, pad), pg.Vector2(WIDTH-pad, HEIGHT-pad))
+        self.data.makeWall(pg.Vector2(WIDTH-pad, HEIGHT-pad), pg.Vector2(pad, HEIGHT-pad))
+        self.data.makeWall(pg.Vector2(pad, HEIGHT-pad), pg.Vector2(pad, pad))
+        self.data.makeWall(pg.Vector2(pad, pad), pg.Vector2(500, 400))
+        self.data.makeWall(pg.Vector2(600, 400), pg.Vector2(WIDTH-pad, HEIGHT-pad))
+        self.data.makeWall(pg.Vector2(500, 150), pg.Vector2(500, 400))
+
+    def mainloop(self):
+        # main loop
+        while True:
+            info = self.stepTime()
+
+            if info == "quit":
+                pg.quit()
+
                 return
+
+    def stepTime(self):
+        if self.render:
+            # Get mouse position if MOUSEFEAR
+            if MOUSEFEAR:
+                mouse_pos = pg.mouse.get_pos()
+                
+                self.data.fears[0] = pg.Vector2(mouse_pos)
+
+            for e in pg.event.get():
+                # Handle quitting
+                if e.type == pg.QUIT or e.type == pg.KEYDOWN and e.key == pg.K_ESCAPE:
+                    return "quit"
+                
+                if e.type == pg.MOUSEBUTTONDOWN and MOUSEFEAR:
+                    # data.fears = np.append(data.fears, [pg.Vector2(mouse_pos)], axis=0)
+                    # data.fears = np.insert(data.fears, pg.Vector2(mouse_pos), len(data.fears))
+                    self.data.fears[self.data.num_fears] = pg.Vector2(mouse_pos)
+                    self.data.num_fears += 1
+
+        dt = self.clock.tick(FPS) / 1000
+
+        self.nBoids.update(dt, TUNING)
+
+        for i in range(len(self.data.fears)):
+            diff = self.data.fear_targets[i] - self.data.fears[i]
+
+            if diff.length() < dt * FEAR_SPEED:
+                self.data.fears[i] = self.data.fear_targets[i]
             
-            if e.type == pg.MOUSEBUTTONDOWN and MOUSEFEAR:
-                data.fears = np.append(data.fears, [[mouse_pos[0], mouse_pos[1], 0]], axis=0)
+            else:
+                self.data.fears[i] = self.data.fears[i] +  diff.normalize() * dt * FEAR_SPEED
 
-        dt = clock.tick(FPS) / 1000
-        screen.fill(BGCOLOR)
-        for fear in data.fears:
-            pg.draw.circle(screen, (255, 0, 0), (fear[0], fear[1]), 5) # Draw red circle on mouse position
-        nBoids.update(dt, TUNING)
-        nBoids.draw(screen)
+        if self.render:
+            # Draw
+            self.screen.fill(BGCOLOR)
+            self.data.drawFears(self.screen)
+            self.data.drawWalls(self.screen)
+            self.nBoids.draw(self.screen)
 
-        if SHOWFPS : screen.blit(font.render(str(int(clock.get_fps())), True, [0,200,0]), (8, 8))
+            if SHOWFPS : self.screen.blit(self.font.render(str(int(self.clock.get_fps())), True, [0,200,0]), (8, 8))
 
-        pg.display.update()
+            pg.display.update()
+        
+        else:
+            print("FPS:", self.clock.get_fps())
+
 
 if __name__ == '__main__':
-    main()  # by Nik
-    pg.quit()
+    sim = Simulation(num_fears=2, num_boids=BOIDZ, render=True)
+
+    # with open("example_infrastructure.json") as f:
+    #     sim.addWallsFromJSON(json.load(f)["walls"][:5])
+
+    sim.data.fear_targets[1] = pg.Vector2(500, 500)
+
+    sim.addTestWalls()
+
+    sim.mainloop()
